+++ agent/cmd/start.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 
 	"github.com/sensu/sensu-go/agent"
-	"github.com/sensu/sensu-go/types"
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/util/path"
 	"github.com/sensu/sensu-go/util/url"
 	"github.com/sensu/sensu-go/version"
@@ -62,6 +62,7 @@ const (
 	flagBackendHandshakeTimeout  = "backend-handshake-timeout"
 	flagBackendHeartbeatInterval = "backend-heartbeat-interval"
 	flagBackendHeartbeatTimeout  = "backend-heartbeat-timeout"
+	flagProtoSerialization       = "protobuf-serialization"
 
 	// TLS flags
 	flagTrustedCAFile         = "trusted-ca-file"
@@ -134,9 +135,10 @@ func newStartCommand(ctx context.Context, args []string, logger *logrus.Entry) *
 			cfg.BackendHandshakeTimeout = viper.GetInt(flagBackendHandshakeTimeout)
 			cfg.BackendHeartbeatInterval = viper.GetInt(flagBackendHeartbeatInterval)
 			cfg.BackendHeartbeatTimeout = viper.GetInt(flagBackendHeartbeatTimeout)
+			cfg.ProtobufSerialization = viper.GetBool(flagProtoSerialization)
 
 			// TLS configuration
-			cfg.TLS = &types.TLSOptions{}
+			cfg.TLS = &corev2.TLSOptions{}
 			cfg.TLS.TrustedCAFile = viper.GetString(flagTrustedCAFile)
 			cfg.TLS.InsecureSkipVerify = viper.GetBool(flagInsecureSkipTLSVerify)
 
@@ -217,10 +219,10 @@ func newStartCommand(ctx context.Context, args []string, logger *logrus.Entry) *
 	viper.SetDefault(flagEventsRateLimit, agent.DefaultEventsAPIRateLimit)
 	viper.SetDefault(flagEventsBurstLimit, agent.DefaultEventsAPIBurstLimit)
 	viper.SetDefault(flagKeepaliveInterval, agent.DefaultKeepaliveInterval)
-	viper.SetDefault(flagKeepaliveTimeout, types.DefaultKeepaliveTimeout)
+	viper.SetDefault(flagKeepaliveTimeout, corev2.DefaultKeepaliveTimeout)
 	viper.SetDefault(flagNamespace, agent.DefaultNamespace)
 	viper.SetDefault(flagPassword, agent.DefaultPassword)
-	viper.SetDefault(flagRedact, types.DefaultRedactFields)
+	viper.SetDefault(flagRedact, corev2.DefaultRedactFields)
 	viper.SetDefault(flagSocketHost, agent.DefaultSocketHost)
 	viper.SetDefault(flagSocketPort, agent.DefaultSocketPort)
 	viper.SetDefault(flagStatsdDisable, agent.DefaultStatsdDisable)
@@ -236,6 +238,7 @@ func newStartCommand(ctx context.Context, args []string, logger *logrus.Entry) *
 	viper.SetDefault(flagBackendHandshakeTimeout, 15)
 	viper.SetDefault(flagBackendHeartbeatInterval, 30)
 	viper.SetDefault(flagBackendHeartbeatTimeout, 45)
+	viper.SetDefault(flagProtoSerialization, false)
 
 	// Merge in config flag set so that it appears in command usage
 	cmd.Flags().AddFlagSet(configFlagSet)
@@ -277,6 +280,7 @@ func newStartCommand(ctx context.Context, args []string, logger *logrus.Entry) *
 	cmd.Flags().Int(flagBackendHandshakeTimeout, viper.GetInt(flagBackendHandshakeTimeout), "number of seconds the agent should wait when negotiating a new WebSocket connection")
 	cmd.Flags().Int(flagBackendHeartbeatInterval, viper.GetInt(flagBackendHeartbeatInterval), "interval at which the agent should send heartbeats to the backend")
 	cmd.Flags().Int(flagBackendHeartbeatTimeout, viper.GetInt(flagBackendHeartbeatTimeout), "number of seconds the agent should wait for a response to a hearbeat")
+	cmd.Flags().Bool(flagProtoSerialization, viper.GetBool(flagProtoSerialization), "enable protobuf serialization/deserialization on the agent websocket handshake")
 
 	cmd.Flags().SetNormalizeFunc(aliasNormalizeFunc(logger))