+++ backend/agentd/session.go
@@ -6,18 +6,15 @@ import (
 	"fmt"
 	"sync"
 
-	jsoniter "github.com/json-iterator/go"
 	"github.com/prometheus/client_golang/prometheus"
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/backend/ringv2"
 	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/handler"
 	"github.com/sensu/sensu-go/transport"
-	"github.com/sensu/sensu-go/types"
 	"github.com/sirupsen/logrus"
 )
-
-var json = jsoniter.ConfigDefault
 
 var (
 	sessionCounter = prometheus.NewGaugeVec(
@@ -28,6 +25,12 @@ var (
 		[]string{"namespace"},
 	)
 )
+
+// MarshalFunc is the function signature for protobuf/JSON marshaling
+type MarshalFunc = func(v interface{}) ([]byte, error)
+
+// UnmarshalFunc is the function signature for protobuf/JSON unmarshaling
+type UnmarshalFunc = func(data []byte, v interface{}) error
 
 // SessionStore specifies the storage requirements of the Session.
 type SessionStore interface {
@@ -53,6 +56,8 @@ type Session struct {
 	ringPool     *ringv2.Pool
 	ctx          context.Context
 	cancel       context.CancelFunc
+	marshal      MarshalFunc
+	unmarshal    UnmarshalFunc
 
 	subscriptions chan messaging.Subscription
 }
@@ -80,7 +85,7 @@ type SessionConfig struct {
 // connection, message bus, and store.
 // The Session is responsible for stopping itself, and does so when it
 // encounters a receive error.
-func NewSession(cfg SessionConfig, conn transport.Transport, bus messaging.MessageBus, store store.Store) (*Session, error) {
+func NewSession(cfg SessionConfig, conn transport.Transport, bus messaging.MessageBus, store store.Store, unmarshal UnmarshalFunc, marshal MarshalFunc) (*Session, error) {
 	// Validate the agent namespace
 	ctx, cancel := context.WithCancel(context.Background())
 	if _, err := store.GetNamespace(ctx, cfg.Namespace); err != nil {
@@ -110,6 +115,8 @@ func NewSession(cfg SessionConfig, conn transport.Transport, bus messaging.Messa
 		ctx:           ctx,
 		cancel:        cancel,
 		ringPool:      cfg.RingPool,
+		unmarshal:     unmarshal,
+		marshal:       marshal,
 	}
 	s.handler = newSessionHandler(s)
 	return s, nil
@@ -173,19 +180,19 @@ func (s *Session) subPump() {
 	for {
 		select {
 		case c := <-s.checkChannel:
-			request, ok := c.(*types.CheckRequest)
+			request, ok := c.(*corev2.CheckRequest)
 			if !ok {
 				logger.Error("session received non-config over check channel")
 				continue
 			}
 
-			configBytes, err := json.Marshal(request)
+			configBytes, err := s.marshal(request)
 			if err != nil {
 				logger.WithError(err).Error("session failed to serialize check request")
 				continue
 			}
 
-			msg := transport.NewMessage(types.CheckRequestType, configBytes)
+			msg := transport.NewMessage(corev2.CheckRequestType, configBytes)
 			s.sendq <- msg
 		case <-s.stopping:
 			return
@@ -202,7 +209,7 @@ func (s *Session) sendPump() {
 	for {
 		select {
 		case msg := <-s.sendq:
-			logger.WithField("message", string(msg.Payload)).Debug("session - sending message")
+			logger.WithField("message_size", len(msg.Payload)).Debug("session - sending message")
 			err := s.conn.Send(msg)
 			if err != nil {
 				switch err := err.(type) {
@@ -286,9 +293,10 @@ func (s *Session) Stop() {
 	}
 }
 
+// handleKeepalive is the keepalive message handler.
 func (s *Session) handleKeepalive(ctx context.Context, payload []byte) error {
-	keepalive := &types.Event{}
-	err := json.Unmarshal(payload, keepalive)
+	keepalive := &corev2.Event{}
+	err := s.unmarshal(payload, keepalive)
 	if err != nil {
 		return err
 	}
@@ -307,10 +315,11 @@ func (s *Session) handleKeepalive(ctx context.Context, payload []byte) error {
 	return s.bus.Publish(messaging.TopicKeepalive, keepalive)
 }
 
+// handleEvent is the event message handler.
 func (s *Session) handleEvent(ctx context.Context, payload []byte) error {
 	// Decode the payload to an event
-	event := &types.Event{}
-	if err := json.Unmarshal(payload, event); err != nil {
+	event := &corev2.Event{}
+	if err := s.unmarshal(payload, event); err != nil {
 		return err
 	}